Add `is.convert` to convert a test to a function
Useful for reuse.

Closes GH-5.
