+++ package.json
@@ -20,7 +20,8 @@
     "Titus Wormer <tituswormer@gmail.com> (https://wooorm.com)"
   ],
   "files": [
-    "index.js"
+    "index.js",
+    "convert.js"
   ],
   "dependencies": {},
   "devDependencies": {
