+++ readme.md
@@ -83,6 +83,14 @@ with `type` set to a non-empty `string`).
 
 `boolean?` — Whether `node` matches.
 
+### `is.convert(test)`
+
+Create a test function from `test`, that can later be called with a `node`,
+`index`, and `parent`.
+Useful if you’re going to test many nodes, for example when creating a utility
+where something else passes an is-compatible test.
+Can also be accessed with `require('unist-util-is/convert')`.
+
 ## Related
 
 *   [`unist-util-find-after`](https://github.com/syntax-tree/unist-util-find-after)
